Move genesis code into its own crate
