+++ Cargo.toml
@@ -34,6 +34,7 @@ name = "lighthouse"
 members = [
 	"beacon_chain/attestation_validation",
 	"beacon_chain/chain",
+	"beacon_chain/genesis",
 	"beacon_chain/naive_fork_choice",
 	"beacon_chain/spec",
 	"beacon_chain/state-transition",
