+++ beacon_chain/genesis/src/lib.rs
@@ -1,0 +1,7 @@
+extern crate spec;
+extern crate types;
+extern crate validator_induction;
+extern crate validator_shuffling;
+
+pub mod beacon_state;
+
